requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
certifi==2020.6.20        # via requests
chardet==3.0.4            # via requests
deprecated==1.2.10        # via pygithub
idna==2.10                # via requests
inflection==0.5.1         # via foliolib (setup.py)
lxml==4.6.1               # via foliolib (setup.py)
psycopg2==2.8.6           # via foliolib (setup.py)
pygithub==1.53            # via foliolib (setup.py)
pyjwt==1.7.1              # via pygithub
python-magic==0.4.18      # via foliolib (setup.py)
pyyaml==5.3.1             # via foliolib (setup.py)
requests==2.24.0          # via pygithub, foliolib (setup.py)
urllib3==1.25.11          # via requests
wrapt==1.12.1             # via deprecated
click==7.1.2
